Move printer to exportable package, update dependencies and play with securing docker images.
