+++ pkg/unittest/results/assertion_result.go
@@ -3,7 +3,7 @@ package results
 import (
 	"fmt"
 
-	"github.com/helm-unittest/helm-unittest/internal/printer"
+	"github.com/helm-unittest/helm-unittest/pkg/unittest/printer"
 )
 
 // AssertionResult result return by Assertion.Assert
