+++ pkg/unittest/results/test_job_result.go
@@ -3,7 +3,7 @@ package results
 import (
 	"time"
 
-	"github.com/helm-unittest/helm-unittest/internal/printer"
+	"github.com/helm-unittest/helm-unittest/pkg/unittest/printer"
 )
 
 // TestJobResult result return by TestJob.Run
