+++ pkg/unittest/utils_test.go
@@ -8,8 +8,8 @@ import (
 	"testing"
 
 	"github.com/helm-unittest/helm-unittest/internal/common"
-	"github.com/helm-unittest/helm-unittest/internal/printer"
 	. "github.com/helm-unittest/helm-unittest/pkg/unittest"
+	"github.com/helm-unittest/helm-unittest/pkg/unittest/printer"
 	"github.com/stretchr/testify/assert"
 )
